+++ src/business/services/transfer.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { TransferEntity, TransferRepository } from '../../data/persistence';
-import { CreateTransferDto, HistoryDto, PaginationDto } from '../../business/dtos';
+import { DateRangeDto, PaginationDto } from '../../business/dtos';
 import { AccountService } from '.';
 
 @Injectable()
@@ -8,47 +8,51 @@ export class TransferService {
 
   constructor(private readonly transferRepository: TransferRepository, private readonly accountService: AccountService) {}
 
-  createTransfer(transfer: CreateTransferDto): TransferEntity {
-    const newTransfer = new TransferEntity();
-    newTransfer.income = this.accountService.getAccountById(transfer.incomeId);
-    newTransfer.outcome = this.accountService.getAccountById(transfer.outcomeId);
-    newTransfer.balance = +transfer.balance;
-    newTransfer.dateTime = transfer.dateTime || Date.now();
-    newTransfer.reason = transfer.reason;
-    return this.transferRepository.register(newTransfer);
+  createTransfer(transfer: TransferEntity): TransferEntity {
+    return this.transferRepository.register(transfer);
   }
 
   getHistoryOut(
     accountId: string,
     pagination?: PaginationDto,
-    dataRange?: HistoryDto,
+    dataRange?: DateRangeDto,
   ): TransferEntity[] {
     if (dataRange){
-      return this.transferRepository.findOutcomeByDataRange(accountId, dataRange.dateInit, dataRange.dateEnd, pagination);
+      return this.transferRepository.findOutcomeByDataRange(accountId, dataRange?.dateInit, dataRange?.dateEnd, pagination);
     } else return this.transferRepository.findByOutcomeId(accountId, pagination);
   }
 
   getHistoryIn(
     accountId: string,
     pagination?: PaginationDto,
-    dataRange?: HistoryDto,
+    dataRange?: DateRangeDto,
   ): TransferEntity[] {
     if (dataRange){
-      return this.transferRepository.findIncomeByDataRange(accountId, dataRange.dateInit, dataRange.dateEnd, pagination);
+      return this.transferRepository.findIncomeByDataRange(accountId, dataRange?.dateInit, dataRange?.dateEnd, pagination);
     } else return this.transferRepository.findByIncomeId(accountId, pagination);
   }
 
   getHistory(
     accountId: string,
     pagination?: PaginationDto,
-    dataRange?: HistoryDto,
+    dataRange?: DateRangeDto,
   ): TransferEntity[] {
     if (dataRange){
-      return this.transferRepository.findIncomeByDataRange(accountId, dataRange.dateInit, dataRange.dateEnd, pagination).concat(this.transferRepository.findOutcomeByDataRange(accountId, dataRange.dateInit, dataRange.dateEnd, pagination));
+      return this.transferRepository.findIncomeByDataRange(accountId, dataRange?.dateInit, dataRange?.dateEnd, pagination).concat(this.transferRepository.findOutcomeByDataRange(accountId, dataRange?.dateInit, dataRange?.dateEnd, pagination));
     } else return this.transferRepository.findByIncomeId(accountId, pagination).concat(this.transferRepository.findByOutcomeId(accountId, pagination));
   }
 
-  deleteTransfer(transferId: string): void {
-    this.transferRepository.delete(transferId);
+  deleteTransfer(transferId: string): boolean {
+    const current = this.transferRepository.findOneById(transferId);
+    if (current){
+      try{
+        this.transferRepository.delete(transferId);
+        return true;
+      } catch {
+        return false;
+      }
+    } else {
+      return false;
+    }
   }
 }
